Neural network implementation with the FANN library
Ours initial approach with the PyBrain library didn't worked very
well, this library is just too slow when training with our database.

So we found this easy to use C library for neural networks that is
much faster than PyBrain.

The code included in this commit still a little ugly (and has a
possible memory leak in the create_network function), but it works
good enough for our needs.