Save the best NN of all epochs
In the previous implementation if the desired error was never achieved
the trainer would keep training the NN until the max number of epochs
was reached, and would save the NN of the last train. The problem with
this behavior was that if over trained the NN started to increase the
mean square error instead decreasing it and the best NN of all epochs
would be lost.

To solve this problem the trainer makes a copy of the NN in each epoch
if it produced a smaller error than all previous epochs, so only the
best of all NNs is saved.